refactor: extract AdminDashboard KPI cards into AdminKPIGrid component
Closes #470
